separete useCategory hook from list
