+++ frontend/src/hooks/use_category.ts
@@ -1,0 +1,22 @@
+import * as React from 'react';
+import { useState, useEffect } from 'react';
+import Client from '../shared/client';
+
+const client = new Client();
+
+const useCategory = () => {
+  const [category, setCategory] = useState<Array<string>>([""]);
+
+  const getCategory = async () => {
+    const json = await client.getCategory();
+    setCategory(json);
+  }
+
+  useEffect(() => {
+    getCategory()
+  }, [])
+
+  return category;
+};
+
+export default useCategory;
